Update SearchBox.js
Since we switched from React to Preact, the app lost the "autofocus" feature on the input when the modal opens.

This happens because [Preact follows the HTML spec](https://github.com/preactjs/preact/issues/1255#issuecomment-439714572) of triggering the autofocus on page load, and not when the component mounts. React is more "aggressive" on this behavior.

This therefore gets rid of the `autoFocus` prop which is ignored here, and focuses the input with an effect.

Closes #119.