chore(webpack): support local network development
This exposes the app to `0.0.0.0` in dev to be able to try the app on mobile on your local network.

I removed the `open` config because (1) it opens it on `0.0.0.0:3000` which displays a "Not secure" message and (2) I find it annoying that it messes up my tabs and virtual desktops.